moving random walk into rmh.py
