+++ blackjax/mcmc/rmh.py
@@ -2,9 +2,8 @@
 from typing import Callable, NamedTuple, Optional, Tuple
 
 import jax
-import jax.numpy as jnp
+from jax import numpy as jnp
 
-import blackjax.mcmc.random_walk as random_walk
 from blackjax.types import Array, PRNGKey, PyTree
 
 __all__ = ["RMHState", "RMHInfo", "init", "kernel"]
@@ -75,7 +74,7 @@ def kernel():
         rng_key: PRNGKey, state: RMHState, logprob_fn: Callable, sigma: Array
     ) -> Tuple[RMHState, RMHInfo]:
 
-        move_proposal_generator = random_walk.normal(sigma)
+        move_proposal_generator = normal(sigma)
 
         def proposal_generator(key_proposal, position):
             move_proposal = move_proposal_generator(key_proposal, position)
@@ -176,3 +175,36 @@ def rmh(
         )
 
     return kernel
+
+
+def normal(sigma: Array) -> Callable:
+    """Normal Random Walk proposal.
+
+    Propose a new position such that its distance to the current position is
+    normally distributed. Suitable for continuous variables.
+
+    Parameter
+    ---------
+    sigma:
+        vector or matrix that contains the standard deviation of the centered
+        normal distribution from which we draw the move proposals.
+
+    """
+    ndim = jnp.ndim(sigma)  # type: ignore[arg-type]
+    shape = jnp.shape(jnp.atleast_1d(sigma))[:1]
+
+    if ndim == 1:
+        dot = jnp.multiply
+    elif ndim == 2:
+        dot = jnp.dot
+    else:
+        raise ValueError
+
+    def propose(rng_key: PRNGKey, position: PyTree) -> PyTree:
+        _, unravel_fn = jax.flatten_util.ravel_pytree(position)
+        sample = jax.random.normal(rng_key, shape)
+        move_sample = dot(sigma, sample)
+        move_unravel = unravel_fn(move_sample)
+        return move_unravel
+
+    return propose
